requirements.txt
pytest-playwright==0.7.1
pytest-xdist==3.8.0
python-dotenv==1.2.1
pytest-order==1.3.0
pytest-html==4.1.1
